requirements.txt
chromadb
google-genai
gunicorn
Flask
